pyproject.toml
[tool.poetry]
name = "wencheng"
version = "0.1.0"
description = "An agent base on GPT to create documents for you"
authors = ["arcstep <43801@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
langserve = "^0.0.46"
langchain-openai = "^0.0.6"
python-dotenv = "^1.0.0"
sse-starlette = "^1.8.2"
pydantic = "1.10.13"
langfuse = "^2.6.0"
uvicorn = "^0.27.0.post1"
dashscope = "^1.14.1"
langchain-chinese = "0.2.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
